+++ rs/ic_os/os_tools/guest_disk/src/generated_key.rs
@@ -13,6 +13,7 @@ const GENERATED_KEY_SIZE_BYTES: usize = 16;
 
 pub struct GeneratedKeyDiskEncryption<'a> {
     pub key_path: &'a Path,
+    pub metrics_file: &'a Path,
 }
 
 impl DiskEncryption for GeneratedKeyDiskEncryption<'_> {
@@ -23,6 +24,11 @@ impl DiskEncryption for GeneratedKeyDiskEncryption<'_> {
             crypt_name,
             &disk_encryption_key,
             activate_flags(partition),
+            // LUKS param verification is not useful for basic disk encryption (outside a trusted
+            // execution environment)
+            /*verify_luks_params=*/
+            false,
+            Some(self.metrics_file),
         )
         .context("Failed to initialize crypt device")?;