+++ rs/sns/test_utils/src/itest_helpers.rs
@@ -352,6 +352,16 @@ pub fn populate_canister_ids(
         swap.nns_governance_canister_id = NNS_GOVERNANCE_CANISTER_ID.to_string();
         swap.icp_ledger_canister_id = ICP_LEDGER_CANISTER_ID.to_string();
     }
+
+    // Index
+    {
+        match sns_canister_init_payloads.index_ng.as_mut() {
+            Some(IndexArg::Init(init_arg)) => {
+                init_arg.ledger_id = ledger_canister_id.0;
+            }
+            other => panic!("bug: expected Some(IndexArg::Init(...)), got {other:?}"),
+        }
+    }
 }
 
 /// All the SNS canisters