+++ packages/pocket-ic/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `PocketIc::update_call_with_effective_principal_and_sender_info`, `PocketIc::update_call_with_sender_info`,
   `PocketIc::query_call_with_effective_principal_and_sender_info`, and `PocketIc::query_call_with_sender_info`
   to make canister calls with sender info (additional information provided by the canister with which the sender principal is associated).
+- The function `PocketIcBuilder::with_test_threshold_keys_subnet` to create a test threshold keys subnet.